Started refactoring NN Strategy
